requirements.txt
langchain-google-genai
python-dotenv
pydantic
tiktoken
playwright
beautifulsoup4
requests
aiohttp
python-socketio 